requirements.txt
matplotlib==3.4.3
numpy==1.21.2
pandas==1.4.2
scipy==1.7.1
mne==0.23.4
scikit-learn~=1.0.1
Pillow==9.1.0
pathlib
msgpack-python
msgpack-numpy==0.4.7.1
mat73
math
joblib==0.14.0
ipywidgets==7.6
multiprocess
torch==1.11.0
torchvision==0.12.0
shutil